@@ -13,28 +13,27 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "doctrine/inflector": "^2.0",
         "nikic/php-parser": "^4.11",
-        "symfony/config": "^5.4.7|^6.0",
-        "symfony/console": "^5.4.7|^6.0",
-        "symfony/dependency-injection": "^5.4.7|^6.0",
+        "symfony/config": "^6.3|^7.0",
+        "symfony/console": "^6.3|^7.0",
+        "symfony/dependency-injection": "^6.3|^7.0",
         "symfony/deprecation-contracts": "^2.2|^3",
-        "symfony/filesystem": "^5.4.7|^6.0",
-        "symfony/finder": "^5.4.3|^6.0",
-        "symfony/framework-bundle": "^5.4.7|^6.0",
-        "symfony/http-kernel": "^5.4.7|^6.0",
-        "symfony/process": "^5.4.7|^6.0"
+        "symfony/filesystem": "^6.3|^7.0",
+        "symfony/finder": "^6.3|^7.0",
+        "symfony/framework-bundle": "^6.3|^7.0",
+        "symfony/http-kernel": "^6.3|^7.0",
+        "symfony/process": "^6.3|^7.0"
     },
     "require-dev": {
         "composer/semver": "^3.0",
-        "doctrine/doctrine-bundle": "^2.4",
+        "doctrine/doctrine-bundle": "^2.5.0",
         "doctrine/orm": "^2.10.0",
-        "symfony/http-client": "^5.4.7|^6.0",
-        "symfony/phpunit-bridge": "^5.4.17|^6.0",
-        "symfony/polyfill-php80": "^1.16.0",
-        "symfony/security-core": "^5.4.7|^6.0",
-        "symfony/yaml": "^5.4.3|^6.0",
+        "symfony/http-client": "^6.3|^7.0",
+        "symfony/phpunit-bridge": "^6.3|^7.0",
+        "symfony/security-core": "^6.3|^7.0",
+        "symfony/yaml": "^6.3|^7.0",
         "twig/twig": "^2.0|^3.0"
     },
     "config": {
@@ -43,8 +42,7 @@
     },
     "conflict": {
         "doctrine/orm": "<2.10",
-        "doctrine/doctrine-bundle": "<2.4",
-        "symfony/doctrine-bridge": "<5.4"
+        "doctrine/doctrine-bundle": "<2.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Bundle\\MakerBundle\\": "src/" }
@@ -54,7 +52,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-main": "1.0-dev"
+            "dev-main": "1.x-dev"
         }
     }
 }
