@@ -2,13 +2,14 @@
 
 namespace App\Entity;
 
-use App\Repository\PublicationFormRepository;
+use App\Repository\PublicationMetaRepository;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Serializer\Annotation\Ignore;
 
 #[
-    ORM\Entity(repositoryClass: PublicationFormRepository::class),
+    ORM\Entity(repositoryClass: PublicationMetaRepository::class),
     ORM\HasLifecycleCallbacks,
     ORM\Table(name: 'publication_meta')
 ]
@@ -18,8 +19,11 @@ class PublicationMeta
         ORM\Id,
         ORM\Column(type: 'bigint', options: ["unsigned" => true])
     ]
-    #[Ignore]
     private $id;
+
+    #[ORM\Column(type: Types::BIGINT, options: ["unsigned" => true])]
+    #[Ignore]
+    private $id_form;
 
     #[ORM\Column(type: 'bigint', options: ["unsigned" => true])]
     #[Ignore]
@@ -30,7 +34,6 @@ class PublicationMeta
     private $id_form_version;
 
     #[ORM\Column(type: 'bigint', options: ["unsigned" => true], nullable: true)]
-    #[Ignore]
     private $id_form_parent;
 
     #[ORM\Column(type: 'string', length: 255, nullable: true)]
@@ -118,6 +121,12 @@ class PublicationMeta
     private $uuid;
 
     #[
+        ORM\ManyToOne(targetEntity: PublicationForm::class, inversedBy: 'publicationMeta', fetch: 'EAGER'),
+        ORM\JoinColumn(name: 'id_form', referencedColumnName: 'id', onDelete: 'CASCADE')
+    ]
+    private $form;
+
+    #[
         ORM\ManyToOne(targetEntity: Publication::class, inversedBy: 'publication_metas', fetch: 'EAGER'),
         ORM\JoinColumn(name: 'id_publication', referencedColumnName: 'id', onDelete: 'CASCADE')
     ]
@@ -162,6 +171,18 @@ class PublicationMeta
 
     public function isId(string $id) : bool {
         return $this->id == $id;
+    }
+
+    public function getIdForm(): ?string
+    {
+        return $this->id_form;
+    }
+
+    public function setIdForm(?string $id_form): static
+    {
+        $this->id_form = $id_form;
+
+        return $this;
     }
 
     #[Ignore]
@@ -419,7 +440,7 @@ class PublicationMeta
         return $this;
     }
 
-    public function getValue(): string
+    public function getValue(): ?string
     {
         return $this->value;
     }
@@ -521,6 +542,19 @@ class PublicationMeta
     }
 
     #[Ignore]
+    public function getForm(): ?PublicationForm
+    {
+        return $this->form;
+    }
+
+    public function setForm(?PublicationForm $form): static
+    {
+        $this->form = $form;
+
+        return $this;
+    }
+
+    #[Ignore]
     public function getPublication(): ?Publication
     {
         return $this->publication;
@@ -532,7 +566,7 @@ class PublicationMeta
 
         return $this;
     }
-    
+
     #[Ignore]
     public function getFormVersion(): ?PublicationFormVersion
     {
