@@ -8,6 +8,7 @@ use App\Entity\PublicationFormVersion;
 use App\Entity\PublicationMeta;
 use App\Entity\PublicationStatus;
 
+use DateTime;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\Collections\Criteria;
@@ -64,83 +65,117 @@ class PublicationService {
         return $sourceData->matching($this->criteria)->first();
     }
 
-    public function getPublicationFormDataById(
+    private function getPublicationFormDataById(
         PersistentCollection | ArrayCollection $sourceData,
         String $id
     ): PublicationForm | FALSE
     {
-        $data = [];
-
         // Criteria or Query
         $this->criteria->where(
             $this->exprBuilder->eq('id', $id)
         );
         
         // FormVersion data
-        $data = $sourceData->matching($this->criteria)->first();
-
-        return $data;
+        return $sourceData->matching($this->criteria)->first();
     }
 
-    public function getPublicationFormDataByUuid(
+    private function getPublicationFormDataByUuid(
         PersistentCollection | ArrayCollection $sourceData,
         String $uuid
     ): PublicationForm | FALSE
     {
-        $data = [];
-
         // Criteria or Query
         $this->criteria->where(
             $this->exprBuilder->eq('uuid', $uuid)
         );
         
         // FormVersion data
-        $data = $sourceData->matching($this->criteria)->first();
-
-        return $data;
+        return $sourceData->matching($this->criteria)->first();
     }
 
-    public function getPublicationMetaDataById(
+    public function getPublicationMetaDataBy(
         PersistentCollection | ArrayCollection $sourceData,
-        String $id
-    ): PublicationMeta | FALSE
+        array $queries
+    ): PublicationMeta | false
     {
-        $data = [];
+        $this->criteria = new Criteria();
 
-        // Criteria or Query
-        $this->criteria->where(
-            $this->exprBuilder->eq('id', $id)
-        );
-        
-        // FormVersion data
-        $data = $sourceData->matching($this->criteria)->first();
+        foreach ($queries as $key => $value) {
+            $this->criteria->where(
+                $this->exprBuilder->eq($key, $value)
+            );
+        }
 
-        return $data;
+        return $sourceData->matching($this->criteria)->first();
     }
 
-    public function getPublicationMetaDataByUuid(
-        PersistentCollection | ArrayCollection $sourceData,
-        String $uuid
-    ): PublicationMeta | FALSE
-    {
-        $data = [];
-
-        // Criteria or Query
-        $this->criteria->where(
-            $this->exprBuilder->eq('uuid', $uuid)
-        );
-        
-        // FormVersion data
-        $data = $sourceData->matching($this->criteria)->first();
-
-        return $data;
-    }
-
-    public function getAllFormMetaData(PersistentCollection $sourceData): Collection
+    public function getAllFormMetaData(
+        PersistentCollection | ArrayCollection $sourceData
+    ): Collection
     {
         $this->criteria = $this->criteria->where(
             $this->exprBuilder->eq('flag_active', true)
         );
+
+        return $sourceData->matching($this->criteria);
+    }
+
+    private function getRequestMetaDataBy(
+        PersistentCollection | ArrayCollection $sourceData,
+        array $queries
+    ): array | false
+    {
+        $this->criteria = new Criteria();
+
+        foreach ($queries as $key => $value) {
+            $this->criteria->where(
+                $this->exprBuilder->eq($key, $value)
+            );
+        }
+
+        return $sourceData->matching($this->criteria)->first();
+    }
+
+    private function getRequestMetaDataByUuid(
+        PersistentCollection | ArrayCollection $sourceData,
+        string $uuid
+    ): array | false
+    {
+        $this->criteria = $this->criteria->where(
+            $this->exprBuilder->eq('uuid', $uuid)
+        );
+
+        return $sourceData->matching($this->criteria)->first();
+    }
+
+    private function getRequestMetaDataByFieldName(
+        PersistentCollection | ArrayCollection $sourceData,
+        ?string $fieldName
+    ): array | false
+    {
+        $this->criteria = $this->criteria->where(
+            $this->exprBuilder->eq('field_name', $fieldName)
+        );
+
+        return $sourceData->matching($this->criteria)->first();
+    }
+
+    private function getFormConfigsByParentId(
+        PersistentCollection | ArrayCollection $sourceData,
+        string $parentId = null
+    ): Collection
+    {
+        $this->criteria = new Criteria();
+
+        if ($parentId) {
+            $this->criteria->where(
+                $this->exprBuilder->eq('id_form_parent', $parentId)
+            );
+        } else {
+            $this->criteria->where(
+                $this->exprBuilder->isNull('id_form_parent')
+            );
+        }
 
         return $sourceData->matching($this->criteria);
     }
@@ -193,25 +228,25 @@ class PublicationService {
 		$formType    			= $formVersion->getPublicationType() ?: null;
 		$formVersion 			= $formVersion ?: null;
 		$formStatus 			= ($this->doctrineManager->getRepository(PublicationStatus::class))->findOneBy([
-			'publication_status_code' => 'DRF'
+			'publication_status_code' => $requestData['publication_status_code']
 		]) ?: null;
-        $title 					=  null;
+        $title 					= null;
 		$publishDate 			= null;
 
         foreach ($requestData['meta_data'] as $metaDataIndex => $metaData) {
             switch ($metaData['field_name']) {
-                case (is_string($titleFieldConfig)) ?
-                    $titleFieldConfig :
-                    $titleFieldConfig->getFieldName():
+                case (is_string($titleFieldConfig))
+                    ? $titleFieldConfig
+                    : $titleFieldConfig->getFieldName():
 
                     $title          = $metaData['value'];
                     break;
 
-                case (is_string($publishDateFieldConfig)) ? 
-                    $publishDateFieldConfig :
-                    $publishDateFieldConfig->getFieldName():
+                case (is_string($publishDateFieldConfig))
+                    ? $publishDateFieldConfig
+                    : $publishDateFieldConfig->getFieldName():
 
-                    $publishDate    = $metaData['value'];
+                    $publishDate    = new DateTime($metaData['value']);
                     break;
 
                 default: break;
@@ -247,6 +282,20 @@ class PublicationService {
 		$results        = $publication;
 		$requestData    = (is_array($request)) ? $request : $request->request->all();
 
+        /**
+         * Remove the old Meta Data if there is Meta Data.
+         * There are two options:
+         * 1. Set active flag to false (0)
+         * 2. Remove existing data from PersistenceCollection 
+         *    array ($results->removePublicationMetas($metaDataConfig);) 
+         */
+        $metaDataConfigs    = $publication->getPublicationMetas();
+        if (count($metaDataConfigs->toArray()) > 0) {
+            foreach ($metaDataConfigs as $metaDataConfigIndex => $metaDataConfig) {
+                $metaDataConfig->setFlagActive(false);
+            }
+        }
+
         // Organize the new Meta Data (create or update)
         if ($request->getMethod() == 'POST') {
             $results    = $this->updateMetaData($requestData, $formVersion, $publication);
@@ -262,81 +311,150 @@ class PublicationService {
     private function updateMetaData(
         array $requestData,
         PublicationFormVersion $formVersion,
-        Publication $publication
+        Publication $publication,
+        PublicationMeta $parentMetaDataConfig = null
     ) : Publication | array
     {
-        $results = $publication;
-        $metaDataConfigs = $publication->getPublicationMetas();
-        $formConfigs = $formVersion->getForms();
+        $results            = $publication;
+        $metaDataConfigs    = $publication->getPublicationMetas();
+        $formConfigs        = $this->getFormConfigsByParentId(
+            $formVersion->getForms(),
+            ($parentMetaDataConfig) ? $parentMetaDataConfig->getIdForm() : null
+        );
 
-        /**
-         * Remove the old Meta Data if there is Meta Data.
-         * There are two options:
-         * 1. Set active flag to false (0)
-         * 2. Remove existing data from PersistenceCollection (array)
-         */
-        if (count($metaDataConfigs->toArray()) > 0) foreach ($metaDataConfigs as $metaDataConfigIndex => $metaDataConfig) {
-            $metaDataConfig->setFlagActive(false);
-            // $results->removePublicationMetas($metaDataConfig);
-        }
+        // Organize data $requestData['meta_data']
+        foreach ($formConfigs->toArray() as $fieldConfigIndex => $fieldConfig) {
 
-        // Organize data
-        foreach ($requestData['meta_data'] as $metaDataIndex => $metaData) {
             /**
              * Initial value:
              * If there is Meta Data in previous Publication Meta Data, then use it as initial value.
              * Other than that, set Meta Data by Form Configuration.
              */
-            $metaDataConfig     = ($this->getPublicationMetaDataByUuid($metaDataConfigs, $metaData['uuid'])) ?
-                $this->getPublicationMetaDataByUuid(
-                    $metaDataConfigs,
-                    $metaData['uuid']
+            $metaData           = ($this->getRequestMetaDataByUuid(
+                new ArrayCollection($requestData['meta_data']),
+                $fieldConfig->getUuid()
+            ))
+                ? $this->getRequestMetaDataByUuid(
+                    new ArrayCollection($requestData['meta_data']),
+                    $fieldConfig->getUuid()
                 )
-            :
-                $this->setPublicationMetaDataByPublicationFormConfig(
+                : $this->getRequestMetaDataByFieldName(
+                    new ArrayCollection($requestData['meta_data']),
+                    $fieldConfig->getFieldName()
+                );
+            
+            $metaDataConfigQueries = ($metaData) ? ['uuid' => $metaData['uuid']] : ['id_form' => $fieldConfig->getId()];
+
+            $metaDataConfig     = (
+                $this->getPublicationMetaDataBy($metaDataConfigs, $metaDataConfigQueries)
+            )
+                ? $this->getPublicationMetaDataBy($metaDataConfigs, $metaDataConfigQueries)
+                : $this->setPublicationMetaDataByPublicationFormConfig(
                     new PublicationMeta(),
                     $publication,
                     $formVersion,
-                    $this->getPublicationFormDataByUuid($formConfigs, $metaData['uuid']),
+                    $fieldConfig,
                     $metaData
-                )
-            ;
+                ); // 4: $this->getPublicationFormDataByUuid($formConfigs, $metaData['uuid'])
 
             $metaDataConfig->setFlagActive(true);
 
             // Specifict handling by type of field
 			switch ($metaDataConfig->getFieldType()) {
-				case 'panel':
-                case 'accordion':
-                case 'well':
-                case 'step':
                 case 'multiple':
-                    /** 
-                     * $metaDataConfig->setValue(
-                     *  $this->getMainDataByDynamicForm(
-                     *      $metaData['value'],
-                     *      $fieldConfig->getChildren()
-                     *  )
-                     * ); 
-                     */
+                    $metaDataConfig->setValue(
+                        $metaData['value'] ?? null
+                    );
+                    dd($metaDataConfig->getFieldType(), $requestData, $metaData);
                     break;
+
+                case 'well':
+                case 'accordion':
+                case 'panel':
+                case 'stepper':
+                case 'step':
+                    $this->updateMetaData(
+                        $requestData,
+                        $formVersion,
+                        $results,
+                        $metaDataConfig
+                    );
+                    break;
+
+                case 'multiple_select':
+                case 'multiple_autoselect':
+                case 'multiple_autocomplete': break;
 
                 case 'select':
                 case 'autoselect':
                 case 'autocomplete':
                     $metaDataConfig->setValue(
-                        $metaData['value']
+                        $metaData['value'] ?? null
+                    );
+                    $metaDataConfig->setOtherValue(
+                        $metaData['other_value'] ?? null
+                    );
+                    break;
+
+                case 'file':
+                case 'image':
+                    $metaDataConfig->setValue(
+                        $metaData['value'] ?? null
                     );
                     $metaDataConfig->setOtherValue([
-                        'text' => $metaData['value'],
-                        'uuid' => $metaData['value'],
+                        'file_name' => null,
+                        'path' => null,
+                        'url' => null
                     ]);
                     break;
 
+                case 'date':
+                case 'month':
+                case 'year':
+                case 'time':
+                case 'datetime':
+                case 'owl-date':
+                case 'owl-month':
+                case 'owl-year':
+                case 'owl-time':
+                case 'owl-datetime':
+                    $metaDataConfig->setValue(
+                        $metaData['value'] ?? null
+                    );
+                    $metaDataConfig->setOtherValue([
+                        'value' => null,
+                        'text' => null
+                    ]);
+                    break;
+
+                case 'daterange':
+                case 'timerange':
+                case 'datetimerange':
+                case 'owl-daterange':
+                case 'owl-timerange':
+                case 'owl-datetimerange': break;
+
+                case 'radio':
+                case 'checkbox':
+                case 'mask':
+                case 'mask_full_time':
+                case 'url':
                 default:
-                    $metaDataConfig->setValue($metaData['value']);
+                    $metaDataConfig->setValue(
+                        $metaData['value'] ?? null
+                    );
                     break;
 			}
+
+            $metaDataConfig->setFlagActive(true);
+
+            /**
+             *  If parent is exist, it`s indicate that cerrent $metaDataConfig is child of a field.
+             *  So, we have to set the parent`s id to map the child is children of parent field.
+             */
+            if ($parentMetaDataConfig) {
+                $metaDataConfig->setIdFormParent($parentMetaDataConfig->getId());
+            }
 
             // Push the Meta Data to Main Data
             $results->addPublicationMetas($metaDataConfig);
@@ -350,7 +468,7 @@ class PublicationService {
         Publication $publication,
         PublicationFormVersion $formVersion,
         PublicationForm $fieldConfig,
-        array $metaData
+        array | false $metaData
     ) : PublicationMeta
     {
         $results = $publicationMeta;
@@ -364,11 +482,9 @@ class PublicationService {
         $results->setUuid($this->commonSvc->createUUID());
 
         // Master data
-        $results->setIdPublication($publication->getId());
+        $results->setForm($fieldConfig);
         $results->setPublication($publication);
-        $results->setIdFormVersion($formVersion->getId());
         $results->setFormVersion($formVersion);
-        $results->setIdFormParent($fieldConfig->getIdFormParent());
 
         // Field configs
         $results->setFieldLabel($fieldConfig->getFieldLabel());
@@ -399,32 +515,69 @@ class PublicationService {
             case 'accordion':
             case 'well':
             case 'step':
-            case 'multiple':
-                /** 
-                 * $results->setValue(
-                 *  $this->getMainDataByDynamicForm(
-                 *      $metaData['value'],
-                 *      $fieldConfig->getChildren()
-                 *  )
-                 * ); 
-                 */
-                break;
+            case 'multiple': break;
+
+            case 'multiple_select':
+            case 'multiple_autoselect':
+            case 'multiple_autocomplete': break;
 
             case 'select':
             case 'autoselect':
             case 'autocomplete':
                 $results->setValue(
-                    $metaData['value']
+                    $metaData['value'] ?? null
+                );
+                $results->setOtherValue(
+                    $metaData['other_value'] ?? null
+                );
+                break;
+
+            case 'file':
+            case 'image':
+                $results->setValue(
+                    $metaData['value'] ?? null
                 );
                 $results->setOtherValue([
-                    'text' => $metaData['value'],
-                    'uuid' => $metaData['value'],
+                    'file_name' => null,
+                    'path' => null,
+                    'url' => null
                 ]);
                 break;
 
+            case 'date':
+            case 'month':
+            case 'year':
+            case 'time':
+            case 'datetime':
+            case 'owl-date':
+            case 'owl-month':
+            case 'owl-year':
+            case 'owl-time':
+            case 'owl-datetime':
+                $results->setValue(
+                    $metaData['value'] ?? null
+                );
+                $results->setOtherValue([
+                    'value' => null,
+                    'text' => null
+                ]);
+                break;
+
+            case 'daterange':
+            case 'timerange':
+            case 'datetimerange':
+            case 'owl-daterange':
+            case 'owl-timerange':
+            case 'owl-datetimerange': break;
+
+            case 'radio':
+            case 'checkbox':
+            case 'mask':
+            case 'mask_full_time':
+            case 'url':
             default:
                 $results->setValue(
-                    $metaData['value']
+                    $metaData['value'] ?? null
                 );
                 break;
         }
@@ -462,7 +615,10 @@ class PublicationService {
 				case 'step':
 				case 'multiple':
 					$results->add([
-						$fieldConfig->getFieldName() => $this->dynamicDataAdjustment($requestData[$fieldConfig->getFieldName()], $fieldConfig->getChildren())
+						$fieldConfig->getFieldName() => $this->dynamicDataAdjustment(
+                            $requestData[$fieldConfig->getFieldName()],
+                            $fieldConfig->getChildren()
+                        )
 					]);
 					break;
 
